backend/requirements.txt
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-GraphQL==2.0.0
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
graphene==2.1.3
graphene-sqlalchemy==2.1.0
mysqlclient==1.3.14
pyjwt==1.4.2
requests==2.21.0

